+++ CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Ability to aim while pressing the shoot button (or other buttons)
+- Added serious bomb button
 - Gyroscope sensibility setting
 - Enabled console
 - Touch controller visibility setting
 - Touch drag sensibility setting
+- Input opacity setting
 
 ## [1.03.3] - 2019-04-20
 ### Fixed